+++ camel/toolkits/medcalc_bench/ideal_body_weight.py
@@ -9,8 +9,8 @@ Modifications include:
 Date: March 2025
 """
 
-from height_conversion import height_conversion_explanation_in
-from rounding import round_number
+from utils.height_conversion import height_conversion_explanation_in
+from utils.rounding import round_number
 
 
 def ibw_explanation(input_variables):
