+++ camel/toolkits/medcalc_bench/utils/height_conversion.py
@@ -8,7 +8,7 @@ Modifications include:
 Date: March 2025
 """
 
-from rounding import round_number
+from .rounding import round_number
 
 
 def height_conversion_explanation(height_info):
