+++ camel/toolkits/medcalc_bench_example/abw_example.py
@@ -30,10 +30,11 @@ def adjusted_body_weight(
     Returns:
         str: A JSON string containing the calculation process and result, formatted as follows:
             {
-                "rationale": "Detailed calculation process and explanatory text",
-                "final_answer": "Adjusted body weight in kilograms (string format)"
+                "weight": ("Weight (float)", "Weight unit (string)"),
+                "height": ("Height (float)", "Height unit (string)"),
+                "sex": "Gender Male/Female",
+                "age": "Age"
             }
-            If an exception occurs, return an error message generated by the `handle_exception` method.
 
     Notes:
         - The `abw_explanation` function is used to calculate the adjusted body weight.
